+++ bindings/wasm/src/route.rs
@@ -84,19 +84,31 @@ pub struct JsRoute {
 #[wasm_bindgen(js_class = Route)]
 impl JsRoute {
     #[wasm_bindgen(getter)]
+    pub fn origin(&self) -> JsValue {
+        let fms = self.inner.borrow();
+        serde_wasm_bindgen::to_value(&fms.route().origin()).unwrap_or_default()
+    }
+
+    #[wasm_bindgen(getter)]
+    pub fn destination(&self) -> JsValue {
+        let fms = self.inner.borrow();
+        serde_wasm_bindgen::to_value(&fms.route().destination()).unwrap_or_default()
+    }
+
+    #[wasm_bindgen(getter)]
     pub fn dist(&self) -> JsValue {
-        let mut fms = self.inner.borrow_mut();
+        let fms = self.inner.borrow();
         serde_wasm_bindgen::to_value(&fms.route().dist()).unwrap_or_default()
     }
 
     #[wasm_bindgen(getter)]
     pub fn ete(&self) -> JsValue {
-        let mut fms = self.inner.borrow_mut();
+        let fms = self.inner.borrow();
         serde_wasm_bindgen::to_value(&fms.route().ete()).unwrap_or_default()
     }
 
     pub fn legs(&self) -> Vec<JsLeg> {
-        let mut fms = self.inner.borrow_mut();
+        let fms = self.inner.borrow();
         fms.route()
             .legs()
             .iter()
@@ -107,7 +119,7 @@ impl JsRoute {
 
     #[wasm_bindgen(js_name = toGeojson)]
     pub fn to_geojson(&self) -> Result<JsValue, JsValue> {
-        let mut fms = self.inner.borrow_mut();
+        let fms = self.inner.borrow();
         let value = fms.route().to_geojson();
         let serializer = Serializer::new().serialize_maps_as_objects(true);
         Ok(value.serialize(&serializer)?)