+++ CHANGELOG.md
@@ -9,15 +9,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Create navigation data per data source
+- Add AIRAC and location to navigation aid
 - Serialization support for navigation data
 - Serialization support for aircraft builder
 - WebAssembly bindings
+
+### Changed
+
+- Modify route and navigation data from closure to reevaluate
+  dependant FMS parts
 
 ### Fixed
 
 - Inclusion of region code in airspace enums
 
+### Removed
+
+- Reading navigation data from file
+
 ## [0.3.3] - 2025-08-13
+
+### Added
 
 - Return the route as GeoJSON
 - Implement std Error trait